project_refactor/requirements.txt
langgraph
langchain-core
langchain-openai
python-dotenv
langgraph-checkpoint-redis
redis
langgraph-checkpoint-sqlite
langchain
langchain-community
google-search-results
faiss-cpu
beautifulsoup4
requests
langchain_tavily
graphviz

#autogen
autogen_agentchat
autogen_ext
autogen_core
tavily-python

#crewai
crewai